Added basic build test and a reusable test function to get valid make arguments.
Added a new test class that just checks that the KeyDaemon can be built,
installed, and executed without errors if given basic valid parameters. Added
a function in testDefs to get these parameters, so they can be reused in both
buildTest and trackedKeyTests. Added options to run individual test modules as
scripts.